requirements.txt
# Dependencias externas
Flask>=2.0
Flask-Login>=0.6.0
Flask-Bcrypt>=1.0.0
Flask-WTF>=1.0.0
Flask-Limiter>=3.5.0
Flask-Caching>=2.1.0
requests>=2.28
Pillow>=9.0
pytesseract>=0.3.10
gunicorn>=21.2.0